providers/requirements.txt
aiohttp
faiss-cpu
gql
langchain
openai
python-dotenv
streamlit
langchain_community
langchain-community
langchain-openai
chromadb
